refactor(field label): flatten structure & make consistent with the lib
