+++ src/FieldLabel/index.js
@@ -1,0 +1,2 @@
+export { Filled as LabelFilled } from './Filled'
+export { Outlined as LabelOutlined } from './Outlined'
