+++ src/InputField/index.js
@@ -4,8 +4,7 @@ import css from 'styled-jsx/css'
 import cx from 'classnames'
 
 import { Input } from './Input'
-import { LabelFilled } from '../forms/LabelFilled'
-import { LabelOutlined } from '../forms/LabelOutlined'
+import { LabelFilled, LabelOutlined } from '../FieldLabel'
 import { colors } from '../theme'
 import { iconStatusPropType, iconStatuses } from '../icons/constants'
 import { inputKinds, inputSizes } from '../forms/constants'
