+++ src/index.js
@@ -5,6 +5,10 @@ export { default as Chip } from './Chip'
 export { default as CircularProgress } from './CircularProgress'
 export { default as Divider } from './Divider'
 export { default as DropdownButton } from './DropdownButton'
+
+// These will be exposed at some point, but are kept internal for now until the API stabilizes
+// export { LabelFilled, LabelOutlined } from './FieldLabel'
+
 export { default as Help } from './Help'
 export { default as InputField } from './InputField'
 export { default as LinearProgress } from './LinearProgress'
